+++ internal/provider/tag_filedistribution_resource.go
@@ -4,14 +4,12 @@ import (
 	"bitbucket.org/booqsoftware/terraform-provider-qbee/internal/qbee"
 	"context"
 	"fmt"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
 
@@ -119,61 +117,11 @@ func (r *tagFiledistributionResource) Schema(_ context.Context, _ resource.Schem
 	}
 }
 
-type FiledistributionTemplate struct {
-	Source      types.String `tfsdk:"source"`
-	Destination types.String `tfsdk:"destination"`
-	IsTemplate  types.Bool   `tfsdk:"is_template"`
-}
-
-func (f FiledistributionTemplate) attrTypes() map[string]attr.Type {
-	return map[string]attr.Type{
-		"source":      types.StringType,
-		"destination": types.StringType,
-		"is_template": types.BoolType,
-	}
-}
-
 type tagFiledistributionResourceModel struct {
 	ID     types.String `tfsdk:"id"`
 	Tag    types.String `tfsdk:"tag"`
 	Extend types.Bool   `tfsdk:"extend"`
 	Files  types.List   `tfsdk:"files"`
-}
-
-type FiledistributionParameter struct {
-	Key   types.String `tfsdk:"key"`
-	Value types.String `tfsdk:"value"`
-}
-
-func (f FiledistributionParameter) attrTypes() map[string]attr.Type {
-	return map[string]attr.Type{
-		"key":   types.StringType,
-		"value": types.StringType,
-	}
-}
-
-type FiledistributionFile struct {
-	Command      types.String `tfsdk:"command"`
-	PreCondition types.String `tfsdk:"pre_condition"`
-	Templates    types.List   `tfsdk:"templates"`
-	Parameters   types.List   `tfsdk:"parameters"`
-}
-
-func (f FiledistributionFile) attrTypes() map[string]attr.Type {
-	return map[string]attr.Type{
-		"command":       types.StringType,
-		"pre_condition": types.StringType,
-		"templates": types.ListType{
-			ElemType: types.ObjectType{
-				AttrTypes: FiledistributionTemplate{}.attrTypes(),
-			},
-		},
-		"parameters": types.ListType{
-			ElemType: types.ObjectType{
-				AttrTypes: FiledistributionParameter{}.attrTypes(),
-			},
-		},
-	}
 }
 
 // Create creates the resource and sets the initial Terraform state.
@@ -241,11 +189,11 @@ func (r *tagFiledistributionResource) writeFiledistribution(ctx context.Context,
 		return fmt.Errorf("%v: %v", diags.Errors()[0].Summary(), diags.Errors()[0].Detail())
 	}
 
-	filesets := r.createFilesets(ctx, files)
+	filesets := PlanToQbeeFilesets(ctx, files)
 
 	// Create the resource
 	tflog.Info(ctx, fmt.Sprintf("Creating file distribution for tag %v with %v filesets", tag, len(filesets)))
-	createResponse, err := r.client.TagConfig.CreateFileDistribution(tag, filesets, extend)
+	createResponse, err := r.client.Config.CreateTagFileDistribution(tag, filesets, extend)
 	if err != nil {
 		return fmt.Errorf("error creating a tag_filedistribution resource: %w", err)
 	}
@@ -265,52 +213,6 @@ func (r *tagFiledistributionResource) writeFiledistribution(ctx context.Context,
 	return nil
 }
 
-func (r *tagFiledistributionResource) createFilesets(ctx context.Context, files []FiledistributionFile) []qbee.FilesetConfig {
-	var filesets []qbee.FilesetConfig
-
-	for _, file := range files {
-		var fsc qbee.FilesetConfig
-
-		if !file.PreCondition.IsNull() {
-			fsc.PreCondition = file.PreCondition.ValueString()
-		}
-
-		if !file.Command.IsNull() {
-			fsc.Command = file.Command.ValueString()
-		}
-
-		var paramValues []FiledistributionParameter
-		file.Parameters.ElementsAs(ctx, &paramValues, false)
-		var params []qbee.FilesetParameter
-		for _, value := range paramValues {
-			params = append(params, qbee.FilesetParameter{
-				Key:   value.Key.ValueString(),
-				Value: value.Value.ValueString(),
-			})
-		}
-
-		var templateValues []FiledistributionTemplate
-		file.Templates.ElementsAs(ctx, &templateValues, false)
-		var templates []qbee.FilesetTemplate
-		for _, value := range templateValues {
-			templates = append(templates, qbee.FilesetTemplate{
-				Source:      value.Source.ValueString(),
-				Destination: value.Destination.ValueString(),
-				IsTemplate:  value.IsTemplate.ValueBool(),
-			})
-		}
-
-		filesets = append(filesets, qbee.FilesetConfig{
-			PreCondition: file.PreCondition.ValueString(),
-			Command:      file.Command.ValueString(),
-			Templates:    templates,
-			Parameters:   params,
-		})
-	}
-
-	return filesets
-}
-
 // Read refreshes the Terraform state with the latest data.
 func (r *tagFiledistributionResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	// Get the current state
@@ -324,7 +226,7 @@ func (r *tagFiledistributionResource) Read(ctx context.Context, req resource.Rea
 	tag := state.Tag.ValueString()
 
 	// Read the real status
-	distribution, err := r.client.TagConfig.GetFiledistribution(tag)
+	distribution, err := r.client.Config.GetTagFiledistribution(tag)
 	if err != nil {
 		resp.Diagnostics.AddError("Could not read tag_filedistribution",
 			"error reading the tag_filedistribution resource: "+err.Error())
@@ -345,66 +247,6 @@ func (r *tagFiledistributionResource) Read(ctx context.Context, req resource.Rea
 	resp.State.Set(ctx, state)
 }
 
-func filedistributionToListValue(ctx context.Context, filedistribution *qbee.GetFileDistributionResponse, resp *resource.ReadResponse) basetypes.ListValue {
-	var files []FiledistributionFile
-
-	for _, file := range filedistribution.Files {
-		files = append(files, FiledistributionFile{
-			Command:      types.StringValue(file.Command),
-			PreCondition: types.StringValue(file.PreCondition),
-			Templates:    templatesToListValue(ctx, file.Templates, resp),
-			Parameters:   parametersToListValue(ctx, file.Parameters, resp),
-		})
-	}
-
-	fileValues, diags := types.ListValueFrom(ctx, types.ObjectType{
-		AttrTypes: FiledistributionFile{}.attrTypes(),
-	}, files)
-	resp.Diagnostics.Append(diags...)
-	return fileValues
-}
-
-func templatesToListValue(ctx context.Context, templates []qbee.FiledistributionFileTemplateResponse, resp *resource.ReadResponse) types.List {
-	if templates == nil {
-		return types.ListNull(basetypes.ObjectType{
-			AttrTypes: FiledistributionTemplate{}.attrTypes(),
-		})
-	}
-
-	var result []FiledistributionTemplate
-	for _, template := range templates {
-		result = append(result, FiledistributionTemplate{
-			Source:      types.StringValue(template.Source),
-			Destination: types.StringValue(template.Destination),
-			IsTemplate:  types.BoolValue(template.IsTemplate),
-		})
-	}
-
-	templatesValue, diags := types.ListValueFrom(ctx, basetypes.ObjectType{AttrTypes: FiledistributionTemplate{}.attrTypes()}, result)
-	resp.Diagnostics.Append(diags...)
-	return templatesValue
-}
-
-func parametersToListValue(ctx context.Context, parameters []qbee.FiledistributionFileParameterResponse, resp *resource.ReadResponse) basetypes.ListValue {
-	if parameters == nil {
-		return types.ListNull(basetypes.ObjectType{
-			AttrTypes: FiledistributionParameter{}.attrTypes(),
-		})
-	}
-
-	var result []FiledistributionParameter
-	for _, parameter := range parameters {
-		result = append(result, FiledistributionParameter{
-			Key:   types.StringValue(parameter.Key),
-			Value: types.StringValue(parameter.Value),
-		})
-	}
-
-	parametersValue, diags := types.ListValueFrom(ctx, basetypes.ObjectType{AttrTypes: FiledistributionParameter{}.attrTypes()}, result)
-	resp.Diagnostics.Append(diags...)
-	return parametersValue
-}
-
 // Delete deletes the resource and removes the Terraform state on success.
 func (r *tagFiledistributionResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	// Retrieve values from the state
@@ -418,7 +260,7 @@ func (r *tagFiledistributionResource) Delete(ctx context.Context, req resource.D
 	// Delete the resource
 	tagName := state.Tag.ValueString()
 	tflog.Info(ctx, fmt.Sprintf("Deleting tag_filedistribution for tag %v", tagName))
-	deleteResponse, err := r.client.TagConfig.ClearFileDistribution(tagName)
+	deleteResponse, err := r.client.Config.ClearTagFileDistribution(tagName)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error deleting tag_filedistribution",